Add files via upload
The file Google Playstore4.csv is a file which is not cleaned i.e. might have missing values, different datatype etc. the other file has all the codes to clean and then visualize the data ,(the path i entered is according to the position of excel file in my laptop).